custom-permissions

+++ D/home/views.py
@@ -5,6 +5,7 @@ from .serializers import PersonSerializer, QuestionSerializer, AnswerSerializer
 from rest_framework.permissions import IsAuthenticated, IsAdminUser, AllowAny
 from django.contrib.auth.models import User
 from rest_framework import status
+from permissions import OwnerOrReadOnly
 
 
 class Home(APIView):
@@ -33,17 +34,24 @@ class QuestionCreateView(APIView):
 
 
 class QuestionUpdateView(APIView):
+    permission_classes = [OwnerOrReadOnly, ]
+
     def put(self, request, pk):
         question = Question.objects.get(pk=pk)
+        self.check_object_permissions(request, question)
         ser_data = QuestionSerializer(instance=question, data=request.data, partial=True)
         if ser_data.is_valid():
-            return Response(ser_data, status=status.HTTP_200_OK)
-        return Response(ser_data, status=status.HTTP_400_BAD_REQUEST)
+            ser_data.save()
+            return Response(ser_data.data, status=status.HTTP_200_OK)
+        return Response(ser_data.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 class QuestionDeleteView(APIView):
+    permission_classes = [OwnerOrReadOnly, ]
+
     def delete(self, request, pk):
         question = Question.objects.get(pk=pk)
+        self.check_object_permissions(request, question)
         question.delete()
         return Response({'message': 'Question Deleted'}, status=status.HTTP_200_OK)
